+++ heylogs-enforcer-rules/src/test/java/nbbrd/heylogs/enforcer/rules/FormatRuleTest.java
@@ -1,0 +1,15 @@
+package nbbrd.heylogs.enforcer.rules;
+
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+class FormatRuleTest {
+
+    @Test
+    void testToString() {
+        FormatRule rule = new FormatRule();
+        assertThat(rule.toString()).contains("FormatChangelogRule").contains("inputFiles");
+    }
+}
+